[EAGLE-986] HBaseStorage unable to deserialize Deserialize from java object bytes
- recover allocatedMB type from int to long

https://issues.apache.org/jira/browse/EAGLE-986

+++ eagle-jpm/eagle-jpm-util/src/main/java/org/apache/eagle/jpm/util/resourcefetch/model/AppInfo.java
@@ -43,7 +43,7 @@ public class AppInfo implements Serializable {
     private long elapsedTime;
     private String amContainerLogs;
     private String amHostHttpAddress;
-    private int allocatedMB;
+    private long allocatedMB;
     private int allocatedVCores;
     private int runningContainers;
     // for HDP 2.7
@@ -186,11 +186,11 @@ public class AppInfo implements Serializable {
         this.amHostHttpAddress = amHostHttpAddress;
     }
 
-    public int getAllocatedMB() {
+    public long getAllocatedMB() {
         return allocatedMB;
     }
 
-    public void setAllocatedMB(int allocatedMB) {
+    public void setAllocatedMB(long allocatedMB) {
         this.allocatedMB = allocatedMB;
     }
